coucou

+++ VM/cmd_live.c
@@ -49,7 +49,7 @@ void			cmd_live(unsigned char *board, t_proc *c_proc, t_cor *core,
 		c_proc->last_lived = 0;
 		if (core->options.verbose == 1)
 		{
-			cmd_verbose_zjmp_live("live", c_proc->pid, vm_num_live);
+			cmd_verbose_zjmp_live("live", c_proc->pid, (int)vm_num_live);
 			write(1, "\n", 1);
 			// ft_printf("P% 5d | live %d\n", c_proc->pid, vm_num_live);
 			cmd_verbose(board, c_proc, 5);
